moved files below 'adapt' files to 'vivliostyle' directory
