+++ src/ts/vivliostyle/columns.ts
@@ -18,11 +18,11 @@
  * @fileoverview Columns - Control column layout.
  */
 import * as Asserts from "./asserts";
-import * as Css from "../adapt/css";
+import * as Css from "./css";
 import * as MathUtil from "./mathutil";
 import * as PageFloats from "./pagefloats";
-import * as Task from "../adapt/task";
-import * as Vtree from "../adapt/vtree";
+import * as Task from "./task";
+import * as Vtree from "./vtree";
 import { Layout } from "./types";
 
 export type ColumnLayoutResult = {
