+++ src/ts/vivliostyle/counters.ts
@@ -18,13 +18,13 @@
  * @fileoverview Counters
  */
 import * as Asserts from "./asserts";
-import * as Base from "../adapt/base";
-import * as CssCasc from "../adapt/csscasc";
-import * as CssProp from "../adapt/cssprop";
-import * as CssStyler from "../adapt/cssstyler";
-import * as Exprs from "../adapt/exprs";
-import * as Vgen from "../adapt/vgen";
-import * as Vtree from "../adapt/vtree";
+import * as Base from "./base";
+import * as CssCasc from "./csscasc";
+import * as CssProp from "./cssprop";
+import * as CssStyler from "./cssstyler";
+import * as Exprs from "./exprs";
+import * as Vgen from "./vgen";
+import * as Vtree from "./vtree";
 import { Layout } from "./types";
 
 /**
