+++ src/ts/vivliostyle/geom.ts
