+++ src/ts/vivliostyle/layoutprocessor.ts
@@ -20,7 +20,7 @@ import * as BreakPosition from "./breakposition";
 import * as Display from "./display";
 import * as LayoutHelper from "./layouthelper";
 import * as Plugin from "./plugin";
-import * as Task from "../adapt/task";
+import * as Task from "./task";
 import { FormattingContextType, Layout, LayoutProcessor, Vtree } from "./types";
 
 /**
