+++ src/ts/vivliostyle/pages.ts
@@ -18,17 +18,17 @@
  * @fileoverview Pages - `@page` rule (CSS Paged Media) support
  */
 import * as Asserts from "./asserts";
-import * as Base from "../adapt/base";
+import * as Base from "./base";
 import * as Constants from "./constants";
-import * as Css from "../adapt/css";
-import * as CssCasc from "../adapt/csscasc";
-import * as CssParse from "../adapt/cssparse";
-import * as CssValid from "../adapt/cssvalid";
-import * as Exprs from "../adapt/exprs";
-import * as Font from "../adapt/font";
-import * as Pm from "../adapt/pm";
+import * as Css from "./css";
+import * as CssCasc from "./csscasc";
+import * as CssParse from "./cssparse";
+import * as CssValid from "./cssvalid";
+import * as Exprs from "./exprs";
+import * as Font from "./font";
+import * as Pm from "./pm";
 import * as Sizing from "./sizing";
-import * as Vtree from "../adapt/vtree";
+import * as Vtree from "./vtree";
 
 /**
  * Resolve page progression direction from writing-mode and direction.
