+++ src/ts/vivliostyle/plugin.ts
@@ -17,11 +17,11 @@
  *
  * @fileoverview Plugin - Plugin mechanism
  */
-import * as Base from "../adapt/base";
-import * as Css from "../adapt/css";
+import * as Base from "./base";
+import * as Css from "./css";
 import * as LayoutProcessor from "./layoutprocessor";
 import * as Logging from "./logging";
-import * as Task from "../adapt/task";
+import * as Task from "./task";
 import { Layout, Vtree } from "./types";
 
 /**
